Type Definition atsamd51g::adc0::inputctrl::MUXPOS_R

type MUXPOS_R = R<u8, MUXPOS_A>;

Reader of field MUXPOS

Implementations

impl MUXPOS_R

pub fn variant(&self) -> MUXPOS_A

Get enumerated values variant

pub fn is_ain0(&self) -> bool

Checks if the value of the field is AIN0

pub fn is_ain1(&self) -> bool

Checks if the value of the field is AIN1

pub fn is_ain2(&self) -> bool

Checks if the value of the field is AIN2

pub fn is_ain3(&self) -> bool

Checks if the value of the field is AIN3

pub fn is_ain4(&self) -> bool

Checks if the value of the field is AIN4

pub fn is_ain5(&self) -> bool

Checks if the value of the field is AIN5

pub fn is_ain6(&self) -> bool

Checks if the value of the field is AIN6

pub fn is_ain7(&self) -> bool

Checks if the value of the field is AIN7

pub fn is_ain8(&self) -> bool

Checks if the value of the field is AIN8

pub fn is_ain9(&self) -> bool

Checks if the value of the field is AIN9

pub fn is_ain10(&self) -> bool

Checks if the value of the field is AIN10

pub fn is_ain11(&self) -> bool

Checks if the value of the field is AIN11

pub fn is_ain12(&self) -> bool

Checks if the value of the field is AIN12

pub fn is_ain13(&self) -> bool

Checks if the value of the field is AIN13

pub fn is_ain14(&self) -> bool

Checks if the value of the field is AIN14

pub fn is_ain15(&self) -> bool

Checks if the value of the field is AIN15

pub fn is_ain16(&self) -> bool

Checks if the value of the field is AIN16

pub fn is_ain17(&self) -> bool

Checks if the value of the field is AIN17

pub fn is_ain18(&self) -> bool

Checks if the value of the field is AIN18

pub fn is_ain19(&self) -> bool

Checks if the value of the field is AIN19

pub fn is_ain20(&self) -> bool

Checks if the value of the field is AIN20

pub fn is_ain21(&self) -> bool

Checks if the value of the field is AIN21

pub fn is_ain22(&self) -> bool

Checks if the value of the field is AIN22

pub fn is_ain23(&self) -> bool

Checks if the value of the field is AIN23

pub fn is_scaledcorevcc(&self) -> bool

Checks if the value of the field is SCALEDCOREVCC

pub fn is_scaledvbat(&self) -> bool

Checks if the value of the field is SCALEDVBAT

pub fn is_scalediovcc(&self) -> bool

Checks if the value of the field is SCALEDIOVCC

pub fn is_bandgap(&self) -> bool

Checks if the value of the field is BANDGAP

pub fn is_ptat(&self) -> bool

Checks if the value of the field is PTAT

pub fn is_ctat(&self) -> bool

Checks if the value of the field is CTAT

pub fn is_dac(&self) -> bool

Checks if the value of the field is DAC

pub fn is_ptc(&self) -> bool

Checks if the value of the field is PTC