Type Definition atsamd51g::adc0::refctrl::REFSEL_R

type REFSEL_R = R<u8, REFSEL_A>;

Reader of field REFSEL

Implementations

impl REFSEL_R

pub fn variant(&self) -> Variant<u8, REFSEL_A>

Get enumerated values variant

pub fn is_intref(&self) -> bool

Checks if the value of the field is INTREF

pub fn is_intvcc0(&self) -> bool

Checks if the value of the field is INTVCC0

pub fn is_intvcc1(&self) -> bool

Checks if the value of the field is INTVCC1

pub fn is_arefa(&self) -> bool

Checks if the value of the field is AREFA

pub fn is_arefb(&self) -> bool

Checks if the value of the field is AREFB

pub fn is_arefc(&self) -> bool

Checks if the value of the field is AREFC