Type Definition atsamd51g::aes::ctrla::AESMODE_R

type AESMODE_R = R<u8, AESMODE_A>;

Reader of field AESMODE

Implementations

impl AESMODE_R

pub fn variant(&self) -> Variant<u8, AESMODE_A>

Get enumerated values variant

pub fn is_ecb(&self) -> bool

Checks if the value of the field is ECB

pub fn is_cbc(&self) -> bool

Checks if the value of the field is CBC

pub fn is_ofb(&self) -> bool

Checks if the value of the field is OFB

pub fn is_cfb(&self) -> bool

Checks if the value of the field is CFB

pub fn is_counter(&self) -> bool

Checks if the value of the field is COUNTER

pub fn is_ccm(&self) -> bool

Checks if the value of the field is CCM

pub fn is_gcm(&self) -> bool

Checks if the value of the field is GCM