Type Definition atsamd51g::dmac::prictrl0::QOS2_R

type QOS2_R = R<u8, QOS2_A>;

Reader of field QOS2

Implementations

impl QOS2_R

pub fn variant(&self) -> QOS2_A

Get enumerated values variant

pub fn is_regular(&self) -> bool

Checks if the value of the field is REGULAR

pub fn is_shortage(&self) -> bool

Checks if the value of the field is SHORTAGE

pub fn is_sensitive(&self) -> bool

Checks if the value of the field is SENSITIVE

pub fn is_critical(&self) -> bool

Checks if the value of the field is CRITICAL