Type Definition atsamd51g::eic::config::SENSE7_R

type SENSE7_R = R<u8, SENSE7_A>;

Reader of field SENSE7

Implementations

impl SENSE7_R

pub fn variant(&self) -> Variant<u8, SENSE7_A>

Get enumerated values variant

pub fn is_none(&self) -> bool

Checks if the value of the field is NONE

pub fn is_rise(&self) -> bool

Checks if the value of the field is RISE

pub fn is_fall(&self) -> bool

Checks if the value of the field is FALL

pub fn is_both(&self) -> bool

Checks if the value of the field is BOTH

pub fn is_high(&self) -> bool

Checks if the value of the field is HIGH

pub fn is_low(&self) -> bool

Checks if the value of the field is LOW