Enum atsamd51g::nvmctrl::ctrlb::CMD_AW

#[repr(u8)]
pub enum CMD_AW {

    EP,
    EB,
    WP,
    WQW,
    SWRST,
    LR,
    UR,
    SPRM,
    CPRM,
    PBC,
    SSB,
    BKSWRST,
    CELCK,
    CEULCK,
    SBPDIS,
    CBPDIS,
    ASEES0,
    ASEES1,
    SEERALOC,
    SEEFLUSH,
    LSEE,
    USEE,
    LSEER,
    USEER,

}

Command

Value on reset: 0

Variants

EP

0: Erase Page - Only supported in the USER and AUX pages.

EB

1: Erase Block - Erases the block addressed by the ADDR register, not supported in the user page

WP

3: Write Page - Writes the contents of the page buffer to the page addressed by the ADDR register, not supported in the user page

WQW

4: Write Quad Word - Writes a 128-bit word at the location addressed by the ADDR register.

SWRST

16: Software Reset - Power-Cycle the NVM memory and replay the device automatic calibration procedure and resets the module configuration registers

LR

17: Lock Region - Locks the region containing the address location in the ADDR register.

UR

18: Unlock Region - Unlocks the region containing the address location in the ADDR register.

SPRM

19: Sets the power reduction mode.

CPRM

20: Clears the power reduction mode.

PBC

21: Page Buffer Clear - Clears the page buffer.

SSB

22: Set Security Bit

BKSWRST

23: Bank swap and system reset, if SMEE is used also reallocate SMEE data into the opposite BANK

CELCK

24: Chip Erase Lock - DSU.CE command is not available

CEULCK

25: Chip Erase Unlock - DSU.CE command is available

SBPDIS

26: Sets STATUS.BPDIS, Boot loader protection is discarded until CBPDIS is issued or next start-up sequence

CBPDIS

27: Clears STATUS.BPDIS, Boot loader protection is not discarded

ASEES0

48: Activate SmartEEPROM Sector 0, deactivate Sector 1

ASEES1

49: Activate SmartEEPROM Sector 1, deactivate Sector 0

SEERALOC

50: Starts SmartEEPROM sector reallocation algorithm

SEEFLUSH

51: Flush SMEE data when in buffered mode

LSEE

52: Lock access to SmartEEPROM data from any mean

USEE

53: Unlock access to SmartEEPROM data

LSEER

54: Lock access to the SmartEEPROM Register Address Space (above 64KB)

USEER

55: Unlock access to the SmartEEPROM Register Address Space (above 64KB)