atsamd51g::nvmctrl::ctrlb

Type Alias CmdW

pub type CmdW<'a, REG> = FieldWriter<'a, REG, 7, Cmdselect>;

Field CMD writer - Command

Aliased Type

struct CmdW<'a, REG> { /* private fields */ }

Implementations

impl<'a, REG> CmdW<'a, REG>
where REG: Writable + RegisterSpec, REG::Ux: From<u8>,

pub fn ep(self) -> &'a mut W<REG>

Erase Page - Only supported in the USER and AUX pages.

pub fn eb(self) -> &'a mut W<REG>

Erase Block - Erases the block addressed by the ADDR register, not supported in the user page

pub fn wp(self) -> &'a mut W<REG>

Write Page - Writes the contents of the page buffer to the page addressed by the ADDR register, not supported in the user page

pub fn wqw(self) -> &'a mut W<REG>

Write Quad Word - Writes a 128-bit word at the location addressed by the ADDR register.

pub fn swrst(self) -> &'a mut W<REG>

Software Reset - Power-Cycle the NVM memory and replay the device automatic calibration procedure and resets the module configuration registers

pub fn lr(self) -> &'a mut W<REG>

Lock Region - Locks the region containing the address location in the ADDR register.

pub fn ur(self) -> &'a mut W<REG>

Unlock Region - Unlocks the region containing the address location in the ADDR register.

pub fn sprm(self) -> &'a mut W<REG>

Sets the power reduction mode.

pub fn cprm(self) -> &'a mut W<REG>

Clears the power reduction mode.

pub fn pbc(self) -> &'a mut W<REG>

Page Buffer Clear - Clears the page buffer.

pub fn ssb(self) -> &'a mut W<REG>

Set Security Bit

pub fn bkswrst(self) -> &'a mut W<REG>

Bank swap and system reset, if SMEE is used also reallocate SMEE data into the opposite BANK

pub fn celck(self) -> &'a mut W<REG>

Chip Erase Lock - DSU.CE command is not available

pub fn ceulck(self) -> &'a mut W<REG>

Chip Erase Unlock - DSU.CE command is available

pub fn sbpdis(self) -> &'a mut W<REG>

Sets STATUS.BPDIS, Boot loader protection is discarded until CBPDIS is issued or next start-up sequence

pub fn cbpdis(self) -> &'a mut W<REG>

Clears STATUS.BPDIS, Boot loader protection is not discarded

pub fn asees0(self) -> &'a mut W<REG>

Activate SmartEEPROM Sector 0, deactivate Sector 1

pub fn asees1(self) -> &'a mut W<REG>

Activate SmartEEPROM Sector 1, deactivate Sector 0

pub fn seeraloc(self) -> &'a mut W<REG>

Starts SmartEEPROM sector reallocation algorithm

pub fn seeflush(self) -> &'a mut W<REG>

Flush SMEE data when in buffered mode

pub fn lsee(self) -> &'a mut W<REG>

Lock access to SmartEEPROM data from any mean

pub fn usee(self) -> &'a mut W<REG>

Unlock access to SmartEEPROM data

pub fn lseer(self) -> &'a mut W<REG>

Lock access to the SmartEEPROM Register Address Space (above 64KB)

pub fn useer(self) -> &'a mut W<REG>

Unlock access to the SmartEEPROM Register Address Space (above 64KB)