Type Definition atsamd51g::nvmctrl::status::BOOTPROT_R

type BOOTPROT_R = R<u8, BOOTPROT_A>;

Reader of field BOOTPROT

Implementations

impl BOOTPROT_R

pub fn variant(&self) -> BOOTPROT_A

Get enumerated values variant

pub fn is_0(&self) -> bool

Checks if the value of the field is _0

pub fn is_8(&self) -> bool

Checks if the value of the field is _8

pub fn is_16(&self) -> bool

Checks if the value of the field is _16

pub fn is_24(&self) -> bool

Checks if the value of the field is _24

pub fn is_32(&self) -> bool

Checks if the value of the field is _32

pub fn is_40(&self) -> bool

Checks if the value of the field is _40

pub fn is_48(&self) -> bool

Checks if the value of the field is _48

pub fn is_56(&self) -> bool

Checks if the value of the field is _56

pub fn is_64(&self) -> bool

Checks if the value of the field is _64

pub fn is_72(&self) -> bool

Checks if the value of the field is _72

pub fn is_80(&self) -> bool

Checks if the value of the field is _80

pub fn is_88(&self) -> bool

Checks if the value of the field is _88

pub fn is_96(&self) -> bool

Checks if the value of the field is _96

pub fn is_104(&self) -> bool

Checks if the value of the field is _104

pub fn is_112(&self) -> bool

Checks if the value of the field is _112

pub fn is_120(&self) -> bool

Checks if the value of the field is _120