Type Definition atsamd51g::sdhc0::psr::BUFWREN_R

type BUFWREN_R = R<bool, BUFWREN_A>;

Reader of field BUFWREN

Implementations

impl BUFWREN_R

pub fn variant(&self) -> BUFWREN_A

Get enumerated values variant

pub fn is_disable(&self) -> bool

Checks if the value of the field is DISABLE

pub fn is_enable(&self) -> bool

Checks if the value of the field is ENABLE