atsamd51g::sercom0::i2cm::status

Type Alias BusstateR

pub type BusstateR = FieldReader<Busstateselect>;

Field BUSSTATE reader - Bus State

Aliased Type

struct BusstateR { /* private fields */ }

Implementations

impl BusstateR

pub const fn variant(&self) -> Busstateselect

Get enumerated values variant

pub fn is_unknown(&self) -> bool

The Bus state is unknown to the I2C Host

pub fn is_idle(&self) -> bool

The Bus state is waiting for a transaction to be initialized

pub fn is_owner(&self) -> bool

The I2C Host is the current owner of the bus

pub fn is_busy(&self) -> bool

Some other I2C Host owns the bus