Type Definition atsamd51g::sercom0::spim::ctrlb::AMODE_R

type AMODE_R = R<u8, AMODE_A>;

Reader of field AMODE

Implementations

impl AMODE_R

pub fn variant(&self) -> Variant<u8, AMODE_A>

Get enumerated values variant

pub fn is_mask(&self) -> bool

Checks if the value of the field is MASK

pub fn is_2_addresses(&self) -> bool

Checks if the value of the field is _2_ADDRESSES

pub fn is_range(&self) -> bool

Checks if the value of the field is RANGE