Type Definition atsamd51g::sercom0::spis::ctrla::CPOL_R

type CPOL_R = R<bool, CPOL_A>;

Reader of field CPOL

Implementations

impl CPOL_R

pub fn variant(&self) -> CPOL_A

Get enumerated values variant

pub fn is_idle_low(&self) -> bool

Checks if the value of the field is IDLE_LOW

pub fn is_idle_high(&self) -> bool

Checks if the value of the field is IDLE_HIGH