Type Definition atsamd51g::sercom0::usart_int::ctrlb::SBMODE_R

type SBMODE_R = R<bool, SBMODE_A>;

Reader of field SBMODE

Implementations

impl SBMODE_R

pub fn variant(&self) -> SBMODE_A

Get enumerated values variant

pub fn is_1_bit(&self) -> bool

Checks if the value of the field is _1_BIT

pub fn is_2_bit(&self) -> bool

Checks if the value of the field is _2_BIT