atsamd51g::system_control
Module actlr

Auxiliary Control Register

Structs

ActlrSpec

Auxiliary Control Register

Type Aliases

DisdefwbufR

Field DISDEFWBUF reader - Disable wruite buffer use during default memory map accesses

DisdefwbufW

Field DISDEFWBUF writer - Disable wruite buffer use during default memory map accesses

DisfoldR

Field DISFOLD reader - Disable IT folding

DisfoldW

Field DISFOLD writer - Disable IT folding

DisfpcaR

Field DISFPCA reader - Disable automatic update of CONTROL.FPCA

DisfpcaW

Field DISFPCA writer - Disable automatic update of CONTROL.FPCA

DismcycintR

Field DISMCYCINT reader - Disable interruption of LDM/STM instructions

DismcycintW

Field DISMCYCINT writer - Disable interruption of LDM/STM instructions

DisoofpR

Field DISOOFP reader - Disable out-of-order FP instructions

DisoofpW

Field DISOOFP writer - Disable out-of-order FP instructions

R

Register ACTLR reader

W

Register ACTLR writer