atsamd51g::tcc0::fctrlb
Type Alias BlankR

pub type BlankR = FieldReader<Blankselect>;

Field BLANK reader - Fault B Blanking Mode

Aliased Type

struct BlankR { /* private fields */ }

Implementations

impl BlankR

pub const fn variant(&self) -> Blankselect

Get enumerated values variant

pub fn is_start(&self) -> bool

Blanking applied from start of the ramp

pub fn is_rise(&self) -> bool

Blanking applied from rising edge of the output waveform

pub fn is_fall(&self) -> bool

Blanking applied from falling edge of the output waveform

pub fn is_both(&self) -> bool

Blanking applied from each toggle of the output waveform